refactor: extract 3 helpers from JsonProjectIo.java (798→333 lines)
Extract package-private helper classes to reduce JsonProjectIo complexity:

- JsonProjectManifest: static manifest read/create, sceneCameraType, version
- JsonTypeResolver: Tweedle type reading, decoding, verification (owns
  TypeReadResult and TypeReadPlan)
- JsonResourceEntryWriter: type/resource entry creation, name sanitization

Inner classes JsonProjectReader/JsonProjectWriter remain as thin orchestrators.
No external API changes — factory methods reader()/writer() unchanged.
All 212 existing tests pass.

Closes #569
